+++ src/app/events/[eventId]/dates/[dailyEventId]/routes/[shuttleRouteId]/edit/components/PriceSectionEarlyBirdType.tsx
@@ -1,0 +1,111 @@
+import { Control, Controller } from 'react-hook-form';
+import Heading from '@/components/text/Heading';
+import NumberInput from '@/components/input/NumberInput';
+import DateInput from '@/components/input/DateInput';
+import { EditFormData } from '../form.type';
+import { discountPercent } from '../../../discountPercent.util';
+
+interface Props {
+  control: Control<EditFormData>;
+  hasEarlybird: boolean;
+  watchRegularPrice: EditFormData['regularPrice'];
+  watchEarlybirdPrice: EditFormData['earlybirdPrice'];
+}
+
+const PriceSectionEarlybirdType = ({
+  control,
+  hasEarlybird,
+  watchRegularPrice,
+  watchEarlybirdPrice,
+}: Props) => {
+  return (
+    <div
+      className={`flex flex-col gap-8 rounded-[4px] p-8 ${hasEarlybird ? '' : 'bg-notion-grey'}`}
+    >
+      <Heading.h5 backgroundColor="blue">얼리버드 가격</Heading.h5>
+      <label className="block text-16 font-500">예약 마감일</label>
+      <Controller
+        control={control}
+        name="earlybirdDeadline"
+        render={({ field: { value } }) => (
+          <DateInput disabled={true} value={value} />
+        )}
+      />
+      <div className="flex flex-col gap-12">
+        <div className="flex flex-col items-start gap-8">
+          <label className="block break-keep text-16 font-500">
+            왕복
+            <span className="ml-4 text-14 text-blue-500">
+              {watchEarlybirdPrice &&
+                discountPercent(
+                  watchRegularPrice.roundTrip,
+                  watchEarlybirdPrice.roundTrip,
+                )}
+            </span>
+          </label>
+          <Controller
+            control={control}
+            name="earlybirdPrice.roundTrip"
+            render={({ field: { onChange, value } }) => (
+              <NumberInput
+                value={value ?? 0}
+                setValue={onChange}
+                disabled={!hasEarlybird}
+              />
+            )}
+          />
+        </div>
+        <div className="flex flex-col items-start gap-8">
+          <label className="block break-keep text-16 font-500">
+            가는편
+            <span className="ml-4 text-14 text-blue-500">
+              {watchEarlybirdPrice &&
+                discountPercent(
+                  watchRegularPrice.toDestination,
+                  watchEarlybirdPrice.toDestination,
+                )}
+            </span>
+          </label>
+          <Controller
+            control={control}
+            name="earlybirdPrice.toDestination"
+            disabled={!hasEarlybird}
+            render={({ field: { onChange, value } }) => (
+              <NumberInput
+                value={value ?? 0}
+                setValue={onChange}
+                disabled={!hasEarlybird}
+              />
+            )}
+          />
+        </div>
+        <div className="flex flex-col items-start gap-8">
+          <label className="block break-keep text-16 font-500">
+            오는편
+            <span className="ml-4 text-14 text-blue-500">
+              {watchEarlybirdPrice &&
+                discountPercent(
+                  watchRegularPrice.fromDestination,
+                  watchEarlybirdPrice.fromDestination,
+                )}
+            </span>
+          </label>
+          <Controller
+            control={control}
+            name="earlybirdPrice.fromDestination"
+            disabled={!hasEarlybird}
+            render={({ field: { onChange, value } }) => (
+              <NumberInput
+                value={value ?? 0}
+                setValue={onChange}
+                disabled={!hasEarlybird}
+              />
+            )}
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default PriceSectionEarlybirdType;
