+++ src/app/events/[eventId]/dates/[dailyEventId]/routes/[shuttleRouteId]/edit/utils/conform.util.ts
@@ -1,10 +1,8 @@
 import { UpdateShuttleRouteRequest } from '@/types/shuttleRoute.type';
-import { UpdateShuttleRouteRequestFormData } from './form.type';
+import { EditFormData } from '../form.type';
 import dayjs from 'dayjs';
 
-export const conform = (
-  data: UpdateShuttleRouteRequestFormData,
-): UpdateShuttleRouteRequest => {
+export const conform = (data: EditFormData): UpdateShuttleRouteRequest => {
   const validateSequenceOrder = (
     hubs: UpdateShuttleRouteRequest['shuttleRouteHubs'],
   ) => {
